do not force libraries onto projects relying on the pkgconfig files
After inspecting the libraries flags added it seems that where
-pthread is passed to build the component -l-pthread is set in the
pkgconfig file which is wrong. This was due to the list adjustment
method used so it either had to be changed or the component
LDFLAGS would had to be ignored completely. I opted for the later
since some vendors (namely Debian) have policies about overlinking
and when packagers start packing projects building against
CopperSpice they will most likely file a report about that anyway.

The CXXFLAGS are required for the WebKit component which is build
with -fPIC and all projects relying on it will have to be build
with that flag so I'm leaving them.

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>