Issue-75: QDockWidget now scaling to fill parent
This was reported on Qt forums as QTBUG-40410
Expanding dockArea does not get additional available space
when using a fixed-size central widget

The issue was marked resolved with the following two commits -

http://code.qt.io/cgit/qt/qtbase.git/commit/?id=2bfee10f3ab8e8e33b2169d735baedcdd9f2b584
QMainWindow: respect the maximum size of the central widget
Make sure the dock widgets get the available space when the
central widget cannot expand any more.
[ChangeLog][QtWidgets][QMainWindow] Dock widgets will now be resized
properly when the central widget has a fixed size.

http://code.qt.io/cgit/qt/qtbase.git/commit/?id=6589fb7e0da291a3d93ff1803d66fb7942bf393e
Workaround for programs depending on previous layout bug
Change 2bfee10f made QDockAreaLayout respect the maximum
size of the central widget. This broke some existing
programs which would set sizePolicy to Fixed on a central
widget without a sizeHint.

This patch restores the old behaviour for the special case
where there are no dock areas. It is highly unlikely that
the intention in this case is to have a very small central
widget with huge blank areas on either side.